Added Validation

+++ src/main/java/com/nagarro/calculator/controllers/CompanyRiskScoreController.java
@@ -1,21 +1,26 @@
 package com.nagarro.calculator.controllers;
 
-//import org.slf4j.Logger;
-//import org.slf4j.LoggerFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 
+
+
 import com.nagarro.calculator.models.CompanyRiskScore;
 import com.nagarro.calculator.services.CompanyRiskScoreService;
+import javax.validation.Valid;
 
 @Controller
 public class CompanyRiskScoreController {
 	
-	//private static final Logger logger = LoggerFactory.getLogger(CompanyRiskScoreController.class);
+	private static final Logger logger = LoggerFactory.getLogger(CompanyRiskScoreController.class);
 	
 	private static final String REDIRECT_PAGE = "redirect:/risk-score";
 	
@@ -30,8 +35,17 @@ public class CompanyRiskScoreController {
 		return "riskScoreAdd";
 	}
 	
-	@PostMapping("/addRiskScore")
-	public String saveRiskScore(@ModelAttribute("companyRiskScore") CompanyRiskScore companyRiskScore) {
+	@PostMapping("/addRiskScore")	
+	public String saveRiskScore(@Valid @ModelAttribute("companyRiskScore") CompanyRiskScore companyRiskScore, BindingResult bindingResult) {
+		
+		if(!companyRiskScoreService.checkDataIfPresent(companyRiskScore)) {
+			bindingResult.addError(new FieldError("companyRiskScore", "companyName", "Company name already present"));
+		}
+		
+		if(bindingResult.hasErrors()) {
+			logger.info("Validations errors while Adding.");
+			return "riskScoreAdd";
+		}
 		 companyRiskScoreService.saveRiskScore(companyRiskScore);
 		 return REDIRECT_PAGE;
 	}
@@ -43,7 +57,16 @@ public class CompanyRiskScoreController {
 	}
 	
 	@PostMapping("/updateRiskScore")
-	public String editRiskScore(@ModelAttribute("companyRiskScore") CompanyRiskScore companyRiskScore) {
+	public String editRiskScore(@Valid @ModelAttribute("companyRiskScore") CompanyRiskScore companyRiskScore, BindingResult bindingResult) {
+		
+		if(companyRiskScoreService.checkDataIfPresent(companyRiskScore)) {
+			bindingResult.addError(new FieldError("companyRiskScore", "companyName", "Company name already present"));
+		}
+		
+		if(bindingResult.hasErrors()) {
+			logger.info("Validations errors while updating.");
+			return "riskScoreUpdate";
+		}
 		
 		CompanyRiskScore existingCompanyRiskScore = companyRiskScoreService.getRiskScoreByName(companyRiskScore);
 		existingCompanyRiskScore.setCompanyName(companyRiskScore.getCompanyName());
@@ -62,7 +85,17 @@ public class CompanyRiskScoreController {
 	}
 	
 	@PostMapping("/deleteRiskScore")
-	public String deleteRiskScore(@ModelAttribute("companyRiskScore") CompanyRiskScore companyRiskScore) {
+	public String deleteRiskScore(@Valid @ModelAttribute("companyRiskScore") CompanyRiskScore companyRiskScore, BindingResult bindingResult) {
+		
+		if(companyRiskScoreService.checkDataIfPresent(companyRiskScore)) {
+			bindingResult.addError(new FieldError("companyRiskScore", "companyName", "Company name already present"));
+		}
+		
+		if(bindingResult.hasErrors()) {
+			logger.info("Validations errors while deleting.");
+			return "riskScoreDelete";
+		}
+		
 		companyRiskScoreService.deleteCompanyRiskScore(companyRiskScore);
 		return REDIRECT_PAGE;
 	}
